refactor: make scripts work without any kludges + add README.md
