+++ package.json
@@ -29,10 +29,10 @@
     "test": "jest",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "debug:tx": "ts-node ./scripts/composeTx.ts",
-    "debug:decomposeTx": "ts-node ./scripts/decomposeTx.ts",
-    "debug:wallet": "ts-node ./scripts/readSecrets.ts",
-    "debug:addViewAccount": "ts-node ./scripts/addViewOnlyAccount.ts"
+    "script:composeTx": "ts-node ./scripts/composeTx.ts",
+    "script:decomposeTx": "ts-node ./scripts/decomposeTx.ts",
+    "script:readSecrets": "ts-node ./scripts/readSecrets.ts",
+    "script:addViewAccount": "ts-node ./scripts/addViewOnlyAccount.ts"
   },
   "repository": {
     "type": "git",
